feat(protocol): 32-bit item ids on the wire (GameU32ItemIds)
Item ids above 65535 need a 32-bit wire field. Add an opt-in GameU32ItemIds
feature that switches every item-id read/write between 16 and 32-bit, defaulting
OFF so this client keeps working with the current u16 server. Both sides MUST
match: enable it only after deploying the 32-bit server build.

- const.h/const.lua: add the GameU32ItemIds feature flag
- InputMessage::getItemId()/OutputMessage::addItemId(): width-switching item-id
  read/write, gated by a static set from the feature at ProtocolGame::onConnect;
  bound to Lua as getItemId/addItemId
- protocolgameparse/protocolgamesend: route every item-id field through
  getItemId()/addItemId() (trades, moves, look, inventory, stash, loot
  containers, supply/impact trackers, item prices, wheel scrolls, proficiency...)
- protocol.lua: use msg:getItemId() for the Lua-parsed item-id reads
- Item/ThingType: widen m_clientId/m_id and their getters to uint32
- features.lua: enable GameU32ItemIds at 1524 (also normalizes this file's
  indentation, which had drifted)

Target: KoliseuOT

+++ src/client/item.h
@@ -130,7 +130,7 @@ public:
     }
     std::vector<ThingType::NpcSaleInfo> getNPCSaleData() { return rawGetThingType()->getNpcSaleData(); }
     uint32 getId() { return m_clientId; }
-    uint16 getClientId() { return m_clientId; }
+    uint32 getClientId() { return m_clientId; }
     uint16 getServerId() { return m_serverId; }
     uint32 getStoreId() { return 0; }
     int getTier() { return m_tier; }
@@ -228,7 +228,7 @@ public:
     bool hasPodiumOutfit() const { return m_podiumOutfit != nullptr; }
 
 private:
-    uint16 m_clientId;
+    uint32 m_clientId;
     uint16 m_serverId;
     uint16 m_countOrSubType;
     stdext::packed_storage<uint8> m_attribs;
